requirements.txt
requests==2.31.0
pytest==8.1.1
pytest-html==4.1.1
pytest-xdist==3.5.0
pydantic==2.6.4
python-dotenv==1.0.1
allure-pytest==2.13.5
